@@ -2,7 +2,7 @@ cmake_minimum_required(VERSION 3.1)
 
 project(crypto)
 
-set(CMAKE_CXX_STANDARD 11)
+set(CMAKE_CXX_STANDARD 14)
 set(CXX_STANDARD_REQUIRED ON)
 
 set(SOURCES
