@@ -1,7 +1,7 @@
-#include "http_connection.h"
 #include "Database.hpp"
 #include "Item.hpp"
 #include "channel_encryption.hpp"
+#include "http_connection.h"
 #include "rate_limiter.h"
 #include "serialization.h"
 #include "server_certificates.h"
@@ -50,8 +50,7 @@ constexpr auto TEST_RETRY_PERIOD = std::chrono::milliseconds(50);
 constexpr size_t MAX_MESSAGE_BODY = 3100;
 
 static void log_error(const error_code& ec) {
-    LOKI_LOG(error) << boost::format("Error(%1%): %2%\n") % ec.value() %
-                           ec.message();
+    LOKI_LOG(error, "Error{}: {}", ec.value(), ec.message());
 }
 
 void make_http_request(boost::asio::io_context& ioc,
@@ -70,8 +69,10 @@ void make_http_request(boost::asio::io_context& ioc,
         resolver.resolve(sn_address, "http", ec);
 #endif
     if (ec) {
-        LOKI_LOG(error) << "http: Failed to parse the IP address. Error code = "
-                        << ec.value() << ". Message: " << ec.message();
+        LOKI_LOG(error,
+                 "http: Failed to parse the IP address. Error code = {}. "
+                 "Message: {}",
+                 ec.value(), ec.message());
         return;
     }
     while (destination != tcp::resolver::iterator()) {
@@ -133,7 +134,7 @@ accept_connection(boost::asio::io_context& ioc,
                   RateLimiter& rate_limiter) {
 
     acceptor.async_accept([&](const error_code& ec, tcp::socket socket) {
-        LOKI_LOG(trace) << "connection accepted";
+        LOKI_LOG(trace, "connection accepted");
         if (!ec)
             std::make_shared<connection_t>(ioc, ssl_ctx, std::move(socket), sn,
                                            channel_encryption, rate_limiter)
@@ -152,7 +153,7 @@ void run(boost::asio::io_context& ioc, std::string& ip, uint16_t port,
          ChannelEncryption<std::string>& channel_encryption,
          RateLimiter& rate_limiter) {
 
-    LOKI_LOG(trace) << "http server run";
+    LOKI_LOG(trace, "http server run");
 
     const auto address =
         boost::asio::ip::make_address(ip); /// throws if incorrect
@@ -181,11 +182,11 @@ connection_t::connection_t(boost::asio::io_context& ioc, ssl::context& ssl_ctx,
       repeat_timer_(ioc), deadline_(ioc, SESSION_TIME_LIMIT),
       notification_ctx_({boost::asio::steady_timer{ioc}, boost::none}) {
 
-    LOKI_LOG(trace) << "connection_t";
+    LOKI_LOG(trace, "connection_t");
     start_timestamp_ = std::chrono::steady_clock::now();
 }
 
-connection_t::~connection_t() { LOKI_LOG(trace) << "~connection_t"; }
+connection_t::~connection_t() { LOKI_LOG(trace, "~connection_t"); }
 
 void connection_t::start() {
     register_deadline();
@@ -202,7 +203,7 @@ void connection_t::do_handshake() {
 
 void connection_t::on_handshake(boost::system::error_code ec) {
     if (ec) {
-        LOKI_LOG(warning) << "ssl handshake failed:" << ec.message();
+        LOKI_LOG(warn, "ssl handshake failed: {}", ec.message());
         return;
     }
 
@@ -210,7 +211,7 @@ void connection_t::on_handshake(boost::system::error_code ec) {
 }
 
 void connection_t::notify(const message_t& msg) {
-    LOKI_LOG(debug) << "Processing message notification: " << msg.data;
+    LOKI_LOG(debug, "Processing message notification: {}", msg.data);
     // save messages, so we can access them once the timer event happens
     notification_ctx_.message = msg;
     // the timer callback will be called once we complete the current callback
@@ -218,7 +219,7 @@ void connection_t::notify(const message_t& msg) {
 }
 
 void connection_t::reset() {
-    LOKI_LOG(debug) << "Resetting the connection";
+    LOKI_LOG(debug, "Resetting the connection");
     notification_ctx_.timer.cancel();
 }
 
@@ -228,7 +229,7 @@ void connection_t::read_request() {
     auto self = shared_from_this();
 
     auto on_data = [self](error_code ec, size_t bytes_transferred) {
-        LOKI_LOG(trace) << "on data: " << bytes_transferred << " bytes";
+        LOKI_LOG(trace, "on data: {} bytes", bytes_transferred);
 
         if (ec) {
             log_error(ec);
@@ -239,7 +240,7 @@ void connection_t::read_request() {
         try {
             self->process_request();
         } catch (const std::exception& e) {
-            LOKI_LOG(error) << "Exception caught: " << e.what();
+            LOKI_LOG(error, "Exception caught: {}", e.what());
             self->body_stream_ << e.what();
         }
 
@@ -254,7 +255,7 @@ void connection_t::read_request() {
 bool connection_t::validate_snode_request() {
     if (!parse_header(LOKI_SENDER_SNODE_PUBKEY_HEADER,
                       LOKI_SNODE_SIGNATURE_HEADER)) {
-        LOKI_LOG(error) << "Missing signature headers";
+        LOKI_LOG(error, "Missing signature headers");
         return false;
     }
     const auto& signature = header_[LOKI_SNODE_SIGNATURE_HEADER];
@@ -264,15 +265,15 @@ bool connection_t::validate_snode_request() {
     const std::string snode_address = public_key_b32z + ".snode";
     if (!service_node_.is_snode_address_known(snode_address)) {
         body_stream_ << "Unknown service node\n";
-        LOKI_LOG(error) << "Discarding signature from unknown service node "
-                        << public_key_b32z;
+        LOKI_LOG(error, "Discarding signature from unknown service node: {}",
+                 public_key_b32z);
         response_.result(http::status::unauthorized);
         return false;
     }
 
     if (!verify_signature(signature, public_key_b32z)) {
         constexpr auto msg = "Could not verify batch signature";
-        LOKI_LOG(warning) << msg;
+        LOKI_LOG(warn, "{}", msg);
         body_stream_ << msg;
         response_.result(http::status::unauthorized);
         return false;
@@ -294,8 +295,7 @@ void connection_t::process_storage_test_req(uint64_t height,
                                             const std::string& tester_addr,
                                             const std::string& msg_hash) {
 
-    LOKI_LOG(trace) << "Performing storage test, attempt: "
-                    << repetition_count_;
+    LOKI_LOG(trace, "Performing storage test, attempt: {}", repetition_count_);
 
     std::string answer;
 
@@ -304,13 +304,11 @@ void connection_t::process_storage_test_req(uint64_t height,
     const auto elapsed_time =
         std::chrono::steady_clock::now() - start_timestamp_;
     if (status == MessageTestStatus::SUCCESS) {
-        LOKI_LOG(debug)
-            << "Storage test success! Attempts: " << repetition_count_
-            << ". Took "
-            << std::chrono::duration_cast<std::chrono::milliseconds>(
-                   elapsed_time)
-                   .count()
-            << "ms";
+        LOKI_LOG(
+            debug, "Storage test success! Attempts: {}. Took {} ms",
+            repetition_count_,
+            std::chrono::duration_cast<std::chrono::milliseconds>(elapsed_time)
+                .count());
         delay_response_ = true;
         body_stream_ << answer;
         response_.result(http::status::ok);
@@ -332,8 +330,8 @@ void connection_t::process_storage_test_req(uint64_t height,
         });
 
     } else {
-        LOKI_LOG(error) << "Failed storage test, tried " << repetition_count_
-                        << " times";
+        LOKI_LOG(error, "Failed storage test, tried {} times.",
+                 repetition_count_);
         response_.result(http::status::bad_request);
         /// TODO: send a helpful error message
     }
@@ -352,14 +350,14 @@ void connection_t::process_swarm_req(boost::string_view target) {
         response_.result(http::status::ok);
         service_node_.process_push_batch(request_.body());
     } else if (target == "v1/swarms/storage_test") {
-        LOKI_LOG(debug) << "Got storage test request";
+        LOKI_LOG(debug, "Got storage test request");
 
         using nlohmann::json;
 
         const json body = json::parse(request_.body(), nullptr, false);
 
         if (body == nlohmann::detail::value_t::discarded) {
-            LOKI_LOG(error) << "Bad snode test request: invalid json";
+            LOKI_LOG(error, "Bad snode test request: invalid json");
             response_.result(http::status::bad_request);
             return;
         }
@@ -372,7 +370,7 @@ void connection_t::process_swarm_req(boost::string_view target) {
             msg_hash = body.at("hash").get<std::string>();
         } catch (...) {
             response_.result(http::status::bad_request);
-            LOKI_LOG(error) << "Bad snode test request: missing fields in json";
+            LOKI_LOG(error, "Bad snode test request: missing fields in json");
             return;
         }
 
@@ -382,17 +380,17 @@ void connection_t::process_swarm_req(boost::string_view target) {
             tester_pk.append(".snode");
             this->process_storage_test_req(blk_height, tester_pk, msg_hash);
         } else {
-            LOKI_LOG(warning) << "Ignoring test request, no pubkey present";
+            LOKI_LOG(warn, "Ignoring test request, no pubkey present");
         }
     } else if (target == "/v1/swarms/blockchain_test") {
-        LOKI_LOG(debug) << "Got blockchain test request";
+        LOKI_LOG(debug, "Got blockchain test request");
 
         using nlohmann::json;
 
         const json body = json::parse(request_.body(), nullptr, false);
 
         if (body.is_discarded()) {
-            LOKI_LOG(error) << "Bad snode test request: invalid json";
+            LOKI_LOG(error, "Bad snode test request: invalid json");
             response_.result(http::status::bad_request);
             return;
         }
@@ -404,7 +402,7 @@ void connection_t::process_swarm_req(boost::string_view target) {
             params.seed = body.at("seed").get<uint64_t>();
         } catch (...) {
             response_.result(http::status::bad_request);
-            LOKI_LOG(error) << "Bad snode test request: missing fields in json";
+            LOKI_LOG(error, "Bad snode test request: missing fields in json");
             return;
         }
 
@@ -423,7 +421,7 @@ void connection_t::process_swarm_req(boost::string_view target) {
         service_node_.perform_blockchain_test(params, callback);
     } else if (target == "/v1/swarms/push") {
 
-        LOKI_LOG(trace) << "swarms/push";
+        LOKI_LOG(trace, "swarms/push");
 
         /// NOTE:: we only expect one message here, but
         /// for now lets reuse the function we already have
@@ -441,7 +439,7 @@ void connection_t::process_request() {
 
     /// This method is responsible for filling out response_
 
-    LOKI_LOG(trace) << "process request";
+    LOKI_LOG(trace, "process request");
     response_.version(request_.version());
     response_.keep_alive(false);
 
@@ -454,15 +452,15 @@ void connection_t::process_request() {
     case http::verb::post:
         if (target == "/v1/storage_rpc") {
             /// Store/load from clients
-            LOKI_LOG(trace) << "got /v1/storage_rpc";
+            LOKI_LOG(trace, "got /v1/storage_rpc");
 
             try {
                 process_client_req();
             } catch (std::exception& e) {
                 response_.result(http::status::internal_server_error);
-                LOKI_LOG(error)
-                    << "exception caught while processing client request: "
-                    << e.what();
+                LOKI_LOG(error,
+                         "exception caught while processing client request: {}",
+                         e.what());
             }
 
             // TODO: parse target (once) to determine if it is a "swarms" call
@@ -486,7 +484,7 @@ void connection_t::process_request() {
         else if (target == "/retrieve_all") {
             process_retrieve_all();
         } else if (target == "/quit") {
-            LOKI_LOG(info) << "got /quit request";
+            LOKI_LOG(info, "got /quit request");
             // a bit of a hack: sending response manually
             delay_response_ = true;
             response_.result(http::status::ok);
@@ -495,7 +493,7 @@ void connection_t::process_request() {
         }
 #endif
         else {
-            LOKI_LOG(error) << "unknown target for POST: " << target;
+            LOKI_LOG(error, "unknown target for POST: {}", target.to_string());
             response_.result(http::status::not_found);
         }
         break;
@@ -504,12 +502,12 @@ void connection_t::process_request() {
         if (target == "/v1/swarms/get_stats") {
             this->on_get_stats();
         } else {
-            LOKI_LOG(error) << "unknown target for GET: " << target;
+            LOKI_LOG(error, "unknown target for GET: {}", target.to_string());
             response_.result(http::status::not_found);
         }
         break;
     default:
-        LOKI_LOG(error) << "bad request";
+        LOKI_LOG(error, "bad request");
         response_.result(http::status::bad_request);
         break;
     }
@@ -539,9 +537,9 @@ void connection_t::write_response() {
             body_stream_ << "Could not encrypt/encode response: ";
             body_stream_ << e.what() << "\n";
             response_.body() = body_stream_.str();
-            LOKI_LOG(error)
-                << "Internal Server Error. Could not encrypt response for "
-                << obfuscate_pubkey(ephemKey);
+            LOKI_LOG(error,
+                     "Internal Server Error. Could not encrypt response for {}",
+                     obfuscate_pubkey(ephemKey));
         }
     }
 #else
@@ -603,8 +601,7 @@ void connection_t::process_store(const json& params) {
             response_.result(http::status::bad_request);
             body_stream_ << boost::format("invalid json: no `%1%` field\n") %
                                 field;
-            LOKI_LOG(error)
-                << boost::format("Bad client request: no `%1%` field") % field;
+            LOKI_LOG(error, "Bad client request: no `{}` field", field);
             return;
         }
     }
@@ -618,7 +615,7 @@ void connection_t::process_store(const json& params) {
     if (pubKey.size() != 66) {
         response_.result(http::status::bad_request);
         body_stream_ << "Pubkey must be 66 characters long\n";
-        LOKI_LOG(error) << "Pubkey must be 66 characters long ";
+        LOKI_LOG(error, "Pubkey must be 66 characters long");
         return;
     }
 
@@ -626,7 +623,7 @@ void connection_t::process_store(const json& params) {
         response_.result(http::status::bad_request);
         body_stream_ << "Message body exceeds maximum allowed length of "
                      << MAX_MESSAGE_BODY << "\n";
-        LOKI_LOG(error) << "Message body too long: " << data.size();
+        LOKI_LOG(error, "Message body too long: {}", data.size());
         return;
     }
 
@@ -636,7 +633,7 @@ void connection_t::process_store(const json& params) {
     }
 
 #ifdef INTEGRATION_TEST
-    LOKI_LOG(trace) << "store body: " << data;
+    LOKI_LOG(trace, "store body: ", data);
 #endif
 
     uint64_t ttlInt;
@@ -644,7 +641,7 @@ void connection_t::process_store(const json& params) {
         response_.result(http::status::forbidden);
         response_.set(http::field::content_type, "text/plain");
         body_stream_ << "Provided TTL is not valid.\n";
-        LOKI_LOG(error) << "Forbidden. Invalid TTL " << ttl;
+        LOKI_LOG(error, "Forbidden. Invalid TTL: {}", ttl);
         return;
     }
     uint64_t timestampInt;
@@ -652,7 +649,7 @@ void connection_t::process_store(const json& params) {
         response_.result(http::status::not_acceptable);
         response_.set(http::field::content_type, "text/plain");
         body_stream_ << "Timestamp error: check your clock\n";
-        LOKI_LOG(error) << "Forbidden. Invalid Timestamp " << timestamp;
+        LOKI_LOG(error, "Forbidden. Invalid Timestamp: {}", timestamp);
         return;
     }
 
@@ -669,7 +666,7 @@ void connection_t::process_store(const json& params) {
 
         json res_body;
         res_body["difficulty"] = service_node_.get_curr_pow_difficulty();
-        LOKI_LOG(error) << "Forbidden. Invalid PoW nonce " << nonce;
+        LOKI_LOG(error, "Forbidden. Invalid PoW nonce: {}", nonce);
 
         /// This might throw if not utf-8 endoded
         body_stream_ << res_body.dump();
@@ -687,8 +684,8 @@ void connection_t::process_store(const json& params) {
         response_.result(http::status::internal_server_error);
         response_.set(http::field::content_type, "text/plain");
         body_stream_ << e.what() << "\n";
-        LOKI_LOG(error) << "Internal Server Error. Could not store message for "
-                        << obfuscate_pubkey(pubKey);
+        LOKI_LOG(error, "Internal Server Error. Could not store message for {}",
+                 obfuscate_pubkey(pubKey));
         return;
     }
 
@@ -696,7 +693,7 @@ void connection_t::process_store(const json& params) {
         response_.result(http::status::service_unavailable);
         response_.set(http::field::content_type, "text/plain");
         body_stream_ << "Service node is initializing\n";
-        LOKI_LOG(warning) << "Service node is initializing";
+        LOKI_LOG(warn, "Service node is initializing");
         return;
     }
 
@@ -705,8 +702,8 @@ void connection_t::process_store(const json& params) {
     json res_body;
     res_body["difficulty"] = service_node_.get_curr_pow_difficulty();
     body_stream_ << res_body.dump();
-    LOKI_LOG(trace) << "Successfully stored message for "
-                    << obfuscate_pubkey(pubKey);
+    LOKI_LOG(trace, "Successfully stored message for {}",
+             obfuscate_pubkey(pubKey));
 }
 
 void connection_t::process_snodes_by_pk(const json& params) {
@@ -714,7 +711,7 @@ void connection_t::process_snodes_by_pk(const json& params) {
     if (!params.contains("pubKey")) {
         response_.result(http::status::bad_request);
         body_stream_ << "invalid json: no `pubKey` field\n";
-        LOKI_LOG(error) << "Bad client request: no `pubKey` field";
+        LOKI_LOG(error, "Bad client request: no `pubKey` field");
         return;
     }
 
@@ -723,7 +720,7 @@ void connection_t::process_snodes_by_pk(const json& params) {
     if (pubKey.size() != 66) {
         response_.result(http::status::bad_request);
         body_stream_ << "Pubkey must be 66 characters long\n";
-        LOKI_LOG(error) << "Pubkey must be 66 characters long ";
+        LOKI_LOG(error, "Pubkey must be 66 characters long ");
         return;
     }
 
@@ -776,7 +773,7 @@ void connection_t::handle_wrong_swarm(const std::string& pubKey) {
 
     /// This might throw if not utf-8 endoded
     body_stream_ << res_body.dump();
-    LOKI_LOG(info) << "Client request for different swarm received";
+    LOKI_LOG(info, "Client request for different swarm received");
 }
 
 constexpr auto LONG_POLL_TIMEOUT = std::chrono::milliseconds(20000);
@@ -813,9 +810,9 @@ void connection_t::poll_db(const std::string& pk,
     if (!service_node_.retrieve(pk, last_hash, items)) {
         response_.result(http::status::internal_server_error);
         response_.set(http::field::content_type, "text/plain");
-        LOKI_LOG(error)
-            << "Internal Server Error. Could not retrieve messages for "
-            << obfuscate_pubkey(pk);
+        LOKI_LOG(error,
+                 "Internal Server Error. Could not retrieve messages for {}",
+                 obfuscate_pubkey(pk));
         return;
     }
 
@@ -823,8 +820,8 @@ void connection_t::poll_db(const std::string& pk,
         request_.find("X-Loki-Long-Poll") != request_.end();
 
     if (!items.empty()) {
-        LOKI_LOG(trace) << "Successfully retrieved messages for "
-                        << obfuscate_pubkey(pk);
+        LOKI_LOG(trace, "Successfully retrieved messages for {}",
+                 obfuscate_pubkey(pk));
     }
 
     if (items.empty() && lp_requested) {
@@ -854,7 +851,7 @@ void connection_t::poll_db(const std::string& pk,
             }
         });
 
-        LOKI_LOG(error) << "just registered notification";
+        LOKI_LOG(error, "just registered notification");
 
     } else {
 
@@ -873,8 +870,7 @@ void connection_t::process_retrieve(const json& params) {
             response_.result(http::status::bad_request);
             body_stream_ << boost::format("invalid json: no `%1%` field\n") %
                                 field;
-            LOKI_LOG(error)
-                << boost::format("Bad client request: no `%1%` field") % field;
+            LOKI_LOG(error, "Bad client request: no `{}` field", field);
             return;
         }
     }
@@ -899,7 +895,7 @@ void connection_t::process_client_req() {
 
 #ifndef DISABLE_ENCRYPTION
     if (!parse_header(LOKI_EPHEMKEY_HEADER)) {
-        LOKI_LOG(error) << "Could not parse headers\n";
+        LOKI_LOG(error, "Could not parse headers");
         return;
     }
 
@@ -913,7 +909,7 @@ void connection_t::process_client_req() {
         response_.set(http::field::content_type, "text/plain");
         body_stream_ << "Could not decode/decrypt body: ";
         body_stream_ << e.what() << "\n";
-        LOKI_LOG(error) << "Bad Request. Could not decrypt body";
+        LOKI_LOG(error, "Bad Request. Could not decrypt body");
         return;
     }
 #endif
@@ -922,7 +918,7 @@ void connection_t::process_client_req() {
     if (body == nlohmann::detail::value_t::discarded) {
         response_.result(http::status::bad_request);
         body_stream_ << "invalid json\n";
-        LOKI_LOG(error) << "Bad client request: invalid json";
+        LOKI_LOG(error, "Bad client request: invalid json");
         return;
     }
 
@@ -930,7 +926,7 @@ void connection_t::process_client_req() {
     if (method_it == body.end() || !method_it->is_string()) {
         response_.result(http::status::bad_request);
         body_stream_ << "invalid json: no `method` field\n";
-        LOKI_LOG(error) << "Bad client request: no method field";
+        LOKI_LOG(error, "Bad client request: no method field");
         return;
     }
 
@@ -940,7 +936,7 @@ void connection_t::process_client_req() {
     if (params_it == body.end() || !params_it->is_object()) {
         response_.result(http::status::bad_request);
         body_stream_ << "invalid json: no `params` field\n";
-        LOKI_LOG(error) << "Bad client request: no params field";
+        LOKI_LOG(error, "Bad client request: no params field");
         return;
     }
 
@@ -953,8 +949,7 @@ void connection_t::process_client_req() {
     } else {
         response_.result(http::status::bad_request);
         body_stream_ << "no method" << method_name << "\n";
-        LOKI_LOG(error) << boost::format("Bad Request. Unknown method '%1%'") %
-                               method_name;
+        LOKI_LOG(error, "Bad Request. Unknown method '{}'", method_name);
     }
 }
 
@@ -971,7 +966,7 @@ void connection_t::register_deadline() {
 
         } else {
 
-            LOKI_LOG(error) << "socket timed out";
+            LOKI_LOG(error, "socket timed out");
             // Close socket to cancel any outstanding operation.
             self->socket_.close(ec);
         }
@@ -991,7 +986,7 @@ void connection_t::on_shutdown(boost::system::error_code ec) {
         ec.assign(0, ec.category());
     }
     if (ec)
-        LOKI_LOG(error) << "Could not close ssl stream gracefully";
+        LOKI_LOG(error, "Could not close ssl stream gracefully");
 
     // At this point the connection is closed gracefully
 }
@@ -1015,7 +1010,7 @@ HttpClientSession::HttpClientSession(boost::asio::io_context& ioc,
 
 void HttpClientSession::on_connect() {
 
-    LOKI_LOG(trace) << "on connect";
+    LOKI_LOG(trace, "on connect");
     http::async_write(socket_, *req_,
                       std::bind(&HttpClientSession::on_write,
                                 shared_from_this(), std::placeholders::_1,
@@ -1024,16 +1019,15 @@ void HttpClientSession::on_connect() {
 
 void HttpClientSession::on_write(error_code ec, size_t bytes_transferred) {
 
-    LOKI_LOG(trace) << "on write";
+    LOKI_LOG(trace, "on write");
     if (ec) {
-        LOKI_LOG(error) << "Error on write, ec: " << ec.value()
-                        << ". Message: " << ec.message();
+        LOKI_LOG(error, "Error on write, ec: {}. Message: {}", ec.value(),
+                 ec.message());
         trigger_callback(SNodeError::ERROR_OTHER, nullptr);
         return;
     }
 
-    LOKI_LOG(trace) << "Successfully transferred " << bytes_transferred
-                    << " bytes";
+    LOKI_LOG(trace, "Successfully transferred {} bytes", bytes_transferred);
 
     // Receive the HTTP response
     http::async_read(socket_, buffer_, res_,
@@ -1043,8 +1037,7 @@ void HttpClientSession::on_write(error_code ec, size_t bytes_transferred) {
 
 void HttpClientSession::on_read(error_code ec, size_t bytes_transferred) {
 
-    LOKI_LOG(trace) << "Successfully received " << bytes_transferred
-                    << " bytes";
+    LOKI_LOG(trace, "Successfully received {} bytes.", bytes_transferred);
 
     if (!ec || (ec == http::error::end_of_stream)) {
 
@@ -1054,16 +1047,16 @@ void HttpClientSession::on_read(error_code ec, size_t bytes_transferred) {
                 std::make_shared<std::string>(res_.body());
             trigger_callback(SNodeError::NO_ERROR, std::move(body));
         } else {
-            LOKI_LOG(error)
-                << "Http request failed, error code: " << res_.result_int();
+            LOKI_LOG(error, "Http request failed, error code: {}",
+                     res_.result_int());
             trigger_callback(SNodeError::HTTP_ERROR, nullptr);
         }
 
     } else {
         /// Do we need to handle `operation aborted` separately here (due to
         /// deadline timer)?
-        LOKI_LOG(error) << "Error on read: " << ec.value()
-                        << ". Message: " << ec.message();
+        LOKI_LOG(error, "Error on read: {}. Message: {}", ec.value(),
+                 ec.message());
         trigger_callback(SNodeError::ERROR_OTHER, nullptr);
     }
 }
@@ -1073,10 +1066,9 @@ void HttpClientSession::start() {
         endpoint_, [this, self = shared_from_this()](const error_code& ec) {
             /// TODO: I think I should just call again if ec == EINTR
             if (ec) {
-                LOKI_LOG(error)
-                    << boost::format(
-                           "Could not connect to %1%, message: %2% (%3%)") %
-                           endpoint_ % ec.message() % ec.value();
+                LOKI_LOG(error, "Could not connect to {}:{}, message: {} ({})",
+                         endpoint_.address().to_string(), endpoint_.port(),
+                         ec.message(), ec.value());
                 trigger_callback(SNodeError::NO_REACH, nullptr);
                 return;
             }
@@ -1092,7 +1084,7 @@ void HttpClientSession::start() {
                     log_error(ec);
                 }
             } else {
-                LOKI_LOG(error) << "client socket timed out";
+                LOKI_LOG(error, "client socket timed out");
                 self->socket_.close();
             }
         });
@@ -1122,8 +1114,7 @@ HttpClientSession::~HttpClientSession() {
     // not_connected happens sometimes so don't bother reporting it.
     if (ec && ec != boost::system::errc::not_connected) {
 
-        LOKI_LOG(error) << "ec: " << ec.value()
-                        << ". Message: " << ec.message();
+        LOKI_LOG(error, "ec: {}. Message: {}", ec.value(), ec.message());
         return;
     }
 }
