@@ -45,7 +45,7 @@ void serialize_message(std::string& res, const T& msg) {
     serialize_integer(res, msg.timestamp);
     serialize(res, msg.nonce);
 
-    LOKI_LOG(trace) << "serialized message: " << msg.data;
+    LOKI_LOG(trace, "serialized message: {}", msg.data);
 }
 
 template void serialize_message(std::string& res, const message_t& msg);
@@ -130,7 +130,7 @@ static boost::optional<uint64_t> deserialize_uint64(string_view& slice) {
 
 std::vector<message_t> deserialize_messages(const std::string& blob) {
 
-    LOKI_LOG(trace) << "=== Deserializing ===";
+    LOKI_LOG(trace, "=== Deserializing ===");
 
     constexpr size_t PK_SIZE = 66; // characters in hex;
 
@@ -143,53 +143,53 @@ std::vector<message_t> deserialize_messages(const std::string& blob) {
         /// Deserialize PK
         auto pk = deserialize_string(slice, PK_SIZE);
         if (!pk) {
-            LOKI_LOG(error) << "could not deserialize pk";
+            LOKI_LOG(error, "Could not deserialize pk");
             return {};
         }
 
         /// Deserialize Hash
         auto hash = deserialize_string(slice);
         if (!hash) {
-            LOKI_LOG(error) << "could not deserialize hash";
+            LOKI_LOG(error, "Could not deserialize hash");
             return {};
         }
 
         /// Deserialize Data
         auto data = deserialize_string(slice);
         if (!data) {
-            LOKI_LOG(error) << "could not deserialize data";
+            LOKI_LOG(error, "Could not deserialize data");
             return {};
         }
 
         /// Deserialize TTL
         auto ttl = deserialize_uint64(slice);
         if (!ttl) {
-            LOKI_LOG(error) << "could not deserialize ttl";
+            LOKI_LOG(error, "Could not deserialize ttl");
             return {};
         }
 
         /// Deserialize Timestamp
         auto timestamp = deserialize_uint64(slice);
         if (!timestamp) {
-            LOKI_LOG(error) << "could not deserialize timestamp";
+            LOKI_LOG(error, "Could not deserialize timestamp");
             return {};
         }
 
         /// Deserialize Nonce
         auto nonce = deserialize_string(slice);
         if (!nonce) {
-            LOKI_LOG(error) << "could not deserialize nonce";
+            LOKI_LOG(error, "Could not deserialize nonce");
             return {};
         }
 
-        LOKI_LOG(trace) << "deserialized data: " << *data;
+        LOKI_LOG(trace, "Deserialized data: {}", *data);
 
-        LOKI_LOG(trace) << boost::format("pk: %1%, msg: %2%") % *pk % *data;
+        LOKI_LOG(trace, "pk: {}, msg: {}", *pk, *data);
 
         result.push_back({*pk, *data, *hash, *ttl, *timestamp, *nonce});
     }
 
-    LOKI_LOG(trace) << "=== END ===";
+    LOKI_LOG(trace, "=== END ===");
 
     return result;
 }
