@@ -82,15 +82,14 @@ void FailedRequestHandler::retry(std::shared_ptr<FailedRequestHandler>&& self) {
 
     attempt_count_ += 1;
     if (attempt_count_ > RETRY_INTERVALS.size()) {
-        LOKI_LOG(debug) << "Gave up after " << attempt_count_ << " attempts";
+        LOKI_LOG(debug, "Gave up after {} attempts", attempt_count_);
         if (give_up_callback_)
             (*give_up_callback_)();
         return;
     }
 
     retry_timer_.expires_after(RETRY_INTERVALS[attempt_count_ - 1]);
-    LOKI_LOG(debug) << "Will retry in "
-                    << RETRY_INTERVALS[attempt_count_ - 1].count() << " secs";
+    LOKI_LOG(debug, "Will retry in {} secs", RETRY_INTERVALS[attempt_count_ - 1].count());
 
     retry_timer_.async_wait([self = std::move(self)](
                                 const boost::system::error_code& ec) mutable {
@@ -105,10 +104,7 @@ void FailedRequestHandler::retry(std::shared_ptr<FailedRequestHandler>&& self) {
         make_sn_request(ioc, sn, req,
                         [self = std::move(self)](sn_response_t&& res) mutable {
                             if (res.error_code != SNodeError::NO_ERROR) {
-                                LOKI_LOG(error)
-                                    << "Could not relay one: " << self->sn_
-                                    << " (attempt #" << self->attempt_count_
-                                    << ")";
+                                LOKI_LOG(error, "Could not relay one: {} (attempt #{})", self->sn_, self->attempt_count_);
                                 self->retry(std::move(self));
                             }
                         });
@@ -116,7 +112,7 @@ void FailedRequestHandler::retry(std::shared_ptr<FailedRequestHandler>&& self) {
 }
 
 FailedRequestHandler::~FailedRequestHandler() {
-    LOKI_LOG(trace) << "~FailedRequestHandler()";
+    LOKI_LOG(trace, "~FailedRequestHandler()");
 }
 
 void FailedRequestHandler::init_timer() { retry(shared_from_this()); }
@@ -205,10 +201,10 @@ ServiceNode::ServiceNode(boost::asio::io_context& ioc,
         throw std::runtime_error("Could not encode our public key");
     }
     // TODO: fail hard if we can't encode our public key
-    LOKI_LOG(info) << "Read our snode address: " << our_address_;
+    LOKI_LOG(info, "Read our snode address: {}", our_address_);
     our_address_.set_port(port);
 
-    LOKI_LOG(info) << "Requesting initial swarm state";
+    LOKI_LOG(info, "Requesting initial swarm state");
     swarm_timer_tick();
     lokid_ping_timer_tick();
     cleanup_timer_tick();
@@ -228,7 +224,7 @@ ServiceNode::~ServiceNode() {
 void ServiceNode::send_sn_request(const std::shared_ptr<request_t>& req,
                                   const sn_record_t& sn) const {
 
-    LOKI_LOG(debug) << "Relaying data to: " << sn;
+    LOKI_LOG(debug, "Relaying data to: {}", sn);
 
     // Note: often one of the reason for failure here is that the node has just
     // deregistered but our SN hasn't updated its swarm list yet.
@@ -237,11 +233,9 @@ void ServiceNode::send_sn_request(const std::shared_ptr<request_t>& req,
             all_stats_.record_request_failed(sn);
 
             if (res.error_code == SNodeError::NO_REACH) {
-                LOKI_LOG(error)
-                    << "Could not relay data to: " << sn << " (Unreachable)";
+                LOKI_LOG(error, "Could not relay data to: {} (Unreachable)", sn);
             } else if (res.error_code == SNodeError::ERROR_OTHER) {
-                LOKI_LOG(error)
-                    << "Could not relay data to: " << sn << " (Generic error)";
+                LOKI_LOG(error, "Could not relay data to: {} (Generic error)", sn);
             }
 
             std::function<void()> give_up_cb = [this, sn]() {
@@ -259,8 +253,7 @@ void ServiceNode::send_sn_request(const std::shared_ptr<request_t>& req,
 void ServiceNode::register_listener(const std::string& pk,
                                     const std::shared_ptr<connection_t>& c) {
     pk_to_listeners[pk].push_back(c);
-    LOKI_LOG(debug) << "register pubkey: " << pk
-                    << ", total pubkeys: " << pk_to_listeners.size();
+    LOKI_LOG(debug, "register pubkey: {}, total pubkeys: {}", pk, pk_to_listeners.size());
 }
 
 void ServiceNode::notify_listeners(const std::string& pk,
@@ -272,7 +265,7 @@ void ServiceNode::notify_listeners(const std::string& pk,
 
         auto& listeners = it->second;
 
-        LOKI_LOG(debug) << "number of notified listeners: " << listeners.size();
+        LOKI_LOG(debug, "number of notified listeners: {}", listeners.size());
 
         for (auto& c : listeners) {
             c->notify(msg);
@@ -305,7 +298,7 @@ void ServiceNode::push_message(const message_t& msg) {
 
     const auto& others = swarm_->other_nodes();
 
-    LOKI_LOG(debug) << "push_message to " << others.size() << " other nodes";
+    LOKI_LOG(debug, "push_message to {} other nodes", others.size());
 
     std::string body;
     serialize_message(body, msg);
@@ -332,7 +325,7 @@ bool ServiceNode::process_store(const message_t& msg) {
 
     /// only accept a message if we are in a swarm
     if (!swarm_) {
-        LOKI_LOG(error) << "error: my swarm in not initialized";
+        LOKI_LOG(error, "error: my swarm in not initialized");
         return false;
     }
 
@@ -362,18 +355,18 @@ void ServiceNode::save_if_new(const message_t& msg) {
     if (db_->store(msg.hash, msg.pub_key, msg.data, msg.ttl, msg.timestamp,
                    msg.nonce)) {
         notify_listeners(msg.pub_key, msg);
-        LOKI_LOG(debug) << "saved message: " << msg.data;
+        LOKI_LOG(debug, "saved message: {}", msg.data);
     }
 }
 
 void ServiceNode::save_bulk(const std::vector<Item>& items) {
 
     if (!db_->bulk_store(items)) {
-        LOKI_LOG(error) << "failed to save batch to the database";
+        LOKI_LOG(error, "failed to save batch to the database");
         return;
     }
 
-    LOKI_LOG(trace) << "saved messages count: " << items.size();
+    LOKI_LOG(trace, "saved messages count: {}", items.size());
 
     // For batches, it is not trivial to get the list of saved (new)
     // messages, so we are only going to "notify" clients with no data
@@ -383,24 +376,21 @@ void ServiceNode::save_bulk(const std::vector<Item>& items) {
 
 void ServiceNode::on_swarm_update(const block_update_t& bu) {
     if (!swarm_) {
-        LOKI_LOG(info) << "Initialized our swarm";
+        LOKI_LOG(info, "Initialized our swarm");
         swarm_ = std::make_unique<Swarm>(our_address_);
     }
 
     if (bu.block_hash != block_hash_) {
 
-        LOKI_LOG(debug) << boost::format("new block, height: %1%, hash: %2%") %
-                               bu.height % bu.block_hash;
+        LOKI_LOG(debug, "new block, height: {}, hash: {}", bu.height, bu.block_hash);
 
         if (bu.height > block_height_ + 1) {
-            LOKI_LOG(warning) << "Skipped some block(s), old: " << block_height_
-                              << " new: " << bu.height;
+            LOKI_LOG(warn, "Skipped some block(s), old: {} new: {}", block_height_, bu.height);
             /// TODO: if we skipped a block, should we try to run peer tests for
             /// them as well?
         } else if (bu.height <= block_height_) {
             // TODO: investigate how testing will be affected under reorg
-            LOKI_LOG(warning)
-                << "new block height is not higher than the current height";
+            LOKI_LOG(warn, "new block height is not higher than the current height");
         }
 
         block_height_ = bu.height;
@@ -409,7 +399,7 @@ void ServiceNode::on_swarm_update(const block_update_t& bu) {
         block_hashes_cache_.push_back(std::make_pair(bu.height, bu.block_hash));
 
     } else {
-        LOKI_LOG(trace) << "already seen this block";
+        LOKI_LOG(trace, "already seen this block");
         return;
     }
 
@@ -447,7 +437,7 @@ parse_swarm_update(const std::shared_ptr<std::string>& response_body,
                    const swarm_callback_t&& cb) {
     const json body = json::parse(*response_body, nullptr, false);
     if (body.is_discarded()) {
-        LOKI_LOG(error) << "Bad lokid rpc response: invalid json";
+        LOKI_LOG(error, "Bad lokid rpc response: invalid json");
         return;
     }
     std::map<swarm_id_t, std::vector<sn_record_t>> swarm_map;
@@ -478,7 +468,7 @@ parse_swarm_update(const std::shared_ptr<std::string>& response_body,
         bu.block_hash = body.at("result").at("block_hash").get<std::string>();
 
     } catch (...) {
-        LOKI_LOG(error) << "Bad lokid rpc response: invalid json";
+        LOKI_LOG(error, "Bad lokid rpc response: invalid json");
         return;
     }
 
@@ -489,7 +479,7 @@ parse_swarm_update(const std::shared_ptr<std::string>& response_body,
     try {
         cb(bu);
     } catch (const std::exception& e) {
-        LOKI_LOG(error) << "Exception caught on swarm update: " << e.what();
+        LOKI_LOG(error, "Exception caught on swarm update: {}", e.what());
     }
 }
 
@@ -497,7 +487,7 @@ void ServiceNode::swarm_timer_tick() {
     const swarm_callback_t cb =
         std::bind(&ServiceNode::on_swarm_update, this, std::placeholders::_1);
 
-    LOKI_LOG(trace) << "UPDATING SWARMS: begin";
+    LOKI_LOG(trace, "UPDATING SWARMS: begin");
 
     json params;
     params["service_node_pubkeys"] = json::array();
@@ -528,9 +518,9 @@ void ServiceNode::lokid_ping_timer_tick() {
 
     auto cb = [](const sn_response_t&& res) {
         if (res.error_code == SNodeError::NO_ERROR) {
-            LOKI_LOG(info) << "Successfully pinged lokid";
+            LOKI_LOG(info, "Successfully pinged lokid");
         } else {
-            LOKI_LOG(warning) << "Could not ping lokid";
+            LOKI_LOG(warn, "Could not ping lokid");
         }
     };
 
@@ -559,7 +549,7 @@ void ServiceNode::perform_blockchain_test(
     bc_test_params_t test_params,
     std::function<void(blockchain_test_answer_t)>&& cb) const {
 
-    LOKI_LOG(debug) << "Delegating blockchain test to lokid";
+    LOKI_LOG(debug, "Delegating blockchain test to lokid");
 
     nlohmann::json params;
 
@@ -568,14 +558,14 @@ void ServiceNode::perform_blockchain_test(
 
     auto on_resp = [cb = std::move(cb)](const sn_response_t& resp) {
         if (resp.error_code != SNodeError::NO_ERROR || !resp.body) {
-            LOKI_LOG(error) << "Could not send blockchain request to Lokid";
+            LOKI_LOG(error, "Could not send blockchain request to Lokid");
             return;
         }
 
         const json body = json::parse(*resp.body, nullptr, false);
 
         if (body.is_discarded()) {
-            LOKI_LOG(error) << "Bad lokid rpc response: invalid json";
+            LOKI_LOG(error, "Bad lokid rpc response: invalid json");
             return;
         }
 
@@ -613,7 +603,7 @@ void ServiceNode::attach_pubkey(std::shared_ptr<request_t>& request) const {
 
 void abort_if_integration_test() {
 #ifdef INTEGRATION_TEST
-    LOKI_LOG(error) << "ABORT in integration test";
+    LOKI_LOG(error, "ABORT in integration test");
     abort();
 #endif
 }
@@ -627,23 +617,19 @@ void ServiceNode::send_storage_test_req(const sn_record_t& testee,
 
         if (res.error_code == SNodeError::NO_ERROR && res.body) {
             if (*res.body == item.data) {
-                LOKI_LOG(debug) << "Storage test is successful for: " << testee
-                                << " at height: " << height;
+                LOKI_LOG(debug, "Storage test is successful for: {} at height: {}", testee, height);
                 success = true;
             } else {
 
-                LOKI_LOG(warning) << "Test answer doesn't match for: " << testee
-                                  << " at height: " << height;
+                LOKI_LOG(warn, "Test answer doesn't match for: {} at height {}", testee, height);
 
 #ifdef INTEGRATION_TEST
-                LOKI_LOG(warning)
-                    << "got: " << *res.body << " expected: " << item.data;
+                LOKI_LOG(warn, "got: {} expected: {}", *res.body, item.data);
 #endif
                 abort_if_integration_test();
             }
         } else {
-            LOKI_LOG(error)
-                << "Failed to send a storage test request to snode: " << testee;
+            LOKI_LOG(error, "Failed to send a storage test request to snode: {}", testee);
 
             /// TODO: retry here, otherwise tests sometimes fail (when SN not
             /// running yet)
@@ -701,8 +687,7 @@ void ServiceNode::process_blockchain_test_response(
     sn_response_t&& res, blockchain_test_answer_t our_answer,
     sn_record_t testee, uint64_t bc_height) {
 
-    LOKI_LOG(debug) << "Processing blockchain test response from: " << testee
-                    << " at height: " << bc_height;
+    LOKI_LOG(debug, "Processing blockchain test response from: {} at height: {}", testee, bc_height);
 
     bool success = false;
 
@@ -715,18 +700,17 @@ void ServiceNode::process_blockchain_test_response(
 
             if (our_answer.res_height == their_height) {
                 success = true;
-                LOKI_LOG(debug) << "Success.";
+                LOKI_LOG(debug, "Success.");
             } else {
-                LOKI_LOG(debug) << "Failed: incorrect answer.";
+                LOKI_LOG(debug, "Failed: incorrect answer.");
             }
 
         } catch (...) {
-            LOKI_LOG(debug) << "Failed: could not find answer in json.";
+            LOKI_LOG(debug, "Failed: could not find answer in json.");
         }
 
     } else {
-        LOKI_LOG(debug) << "Failed to send a blockchain test request to snode: "
-                        << testee;
+        LOKI_LOG(debug, "Failed to send a blockchain test request to snode: {}", testee);
     }
 
     this->all_stats_.record_blockchain_test_result(testee, success);
@@ -740,7 +724,7 @@ bool ServiceNode::derive_tester_testee(uint64_t blk_height, sn_record_t& tester,
     members.push_back(our_address_);
 
     if (members.size() < 2) {
-        LOKI_LOG(error) << "Could not initiate peer test: swarm too small";
+        LOKI_LOG(error, "Could not initiate peer test: swarm too small");
         return false;
     }
 
@@ -751,7 +735,7 @@ bool ServiceNode::derive_tester_testee(uint64_t blk_height, sn_record_t& tester,
         block_hash = block_hash_;
     } else if (blk_height < block_height_) {
 
-        LOKI_LOG(debug) << "got storage test request for an older block";
+        LOKI_LOG(debug, "got storage test request for an older block");
 
         const auto it =
             std::find_if(block_hashes_cache_.begin(), block_hashes_cache_.end(),
@@ -762,19 +746,19 @@ bool ServiceNode::derive_tester_testee(uint64_t blk_height, sn_record_t& tester,
         if (it != block_hashes_cache_.end()) {
             block_hash = it->second;
         } else {
-            LOKI_LOG(warning) << "Could not find hash for a given block height";
+            LOKI_LOG(warn, "Could not find hash for a given block height");
             // TODO: request from lokid?
             return false;
         }
     } else {
         assert(false);
-        LOKI_LOG(error) << "Could not find hash: block height is in the future";
+        LOKI_LOG(error, "Could not find hash: block height is in the future");
         return false;
     }
 
     uint64_t seed;
     if (block_hash.size() < sizeof(seed)) {
-        LOKI_LOG(error) << "Could not initiate peer test: invalid block hash";
+        LOKI_LOG(error, "Could not initiate peer test: invalid block hash");
         return false;
     }
 
@@ -802,8 +786,7 @@ MessageTestStatus ServiceNode::process_storage_test_req(
     std::string block_hash;
 
     if (blk_height > block_height_) {
-        LOKI_LOG(warning) << "Our blockchain is behind, height: "
-                          << block_height_ << ", requested: " << blk_height;
+        LOKI_LOG(warn, "Our blockchain is behind, height: {}, requested: {}", block_height_, blk_height);
         return MessageTestStatus::RETRY;
     }
 
@@ -814,18 +797,16 @@ MessageTestStatus ServiceNode::process_storage_test_req(
         derive_tester_testee(blk_height, tester, testee);
 
         if (testee != our_address_) {
-            LOKI_LOG(warning)
-                << "We are NOT the testee for height: " << blk_height;
+            LOKI_LOG(warn, "We are NOT the testee for height: {}", blk_height);
             return MessageTestStatus::ERROR;
         }
 
         if (tester.sn_address() != tester_addr) {
-            LOKI_LOG(warning) << "Wrong tester: " << tester_addr
-                              << ", expected: " << tester.sn_address();
+            LOKI_LOG(warn, "Wrong tester: {}, expected: {}", tester_addr, tester.sn_address());
             abort_if_integration_test();
             return MessageTestStatus::ERROR;
         } else {
-            LOKI_LOG(trace) << "Tester is valid: " << tester_addr;
+            LOKI_LOG(trace, "Tester is valid: {}", tester_addr);
         }
     }
 
@@ -843,15 +824,14 @@ bool ServiceNode::select_random_message(Item& item) {
 
     uint64_t message_count;
     if (!db_->get_message_count(message_count)) {
-        LOKI_LOG(error) << "could not count messages in the database";
+        LOKI_LOG(error, "Could not count messages in the database");
         return false;
     }
 
-    LOKI_LOG(info) << "total messages: " << message_count;
+    LOKI_LOG(info, "total messages: {}", message_count);
 
     if (message_count == 0) {
-        LOKI_LOG(warning)
-            << "no messages in the database to initiate a peer test";
+        LOKI_LOG(warn, "No messages in the database to initiate a peer test");
         return false;
     }
 
@@ -863,7 +843,7 @@ bool ServiceNode::select_random_message(Item& item) {
     const auto msg_idx = util::uniform_distribution_portable(mt, message_count);
 
     if (!db_->retrieve_by_index(msg_idx, item)) {
-        LOKI_LOG(error) << "could not retrieve message by index: " << msg_idx;
+        LOKI_LOG(error, "Could not retrieve message by index: {}", msg_idx);
         return false;
     }
 
@@ -878,8 +858,7 @@ void ServiceNode::initiate_peer_test() {
         return;
     }
 
-    LOKI_LOG(trace) << "For height " << block_height_ << " the tester is "
-                    << tester << " testee: " << testee;
+    LOKI_LOG(trace, "For height {}; tester: {} testee: {}", block_height_, tester, testee);
 
     if (tester != our_address_) {
         /// Not our turn to initiate a test
@@ -891,10 +870,9 @@ void ServiceNode::initiate_peer_test() {
         // 2.1. Select a message
         Item item;
         if (!this->select_random_message(item)) {
-            LOKI_LOG(error) << "Could not select a message for testing";
+            LOKI_LOG(error, "Could not select a message for testing");
         } else {
-            LOKI_LOG(trace) << "selected random message : " << item.hash << ", "
-                            << item.data;
+            LOKI_LOG(trace, "Selected random message: {}, {}", item.hash, item.data);
 
             // 2.2. Initiate testing request
             send_storage_test_req(testee, item);
@@ -915,8 +893,7 @@ void ServiceNode::initiate_peer_test() {
         constexpr uint64_t SAFETY_BUFFER_BLOCKS = CHECKPOINT_DISTANCE * 2;
 
         if (block_height_ <= SAFETY_BUFFER_BLOCKS) {
-            LOKI_LOG(debug)
-                << "Blockchain too short, skipping blockchain testing.";
+            LOKI_LOG(debug, "Blockchain too short, skipping blockchain testing.");
             return;
         }
 
@@ -969,16 +946,16 @@ void ServiceNode::bootstrap_swarms(
     const std::vector<swarm_id_t>& swarms) const {
 
     if (swarms.empty()) {
-        LOKI_LOG(info) << "bootstrapping all swarms\n";
+        LOKI_LOG(info, "Bootstrapping all swarms");
     } else {
-        LOKI_LOG(info) << "bootstrapping swarms: " << vec_to_string(swarms);
+        LOKI_LOG(info, "Bootstrapping swarms: {}", vec_to_string(swarms));
     }
 
     const auto& all_swarms = swarm_->all_swarms();
 
     std::vector<Item> all_entries;
     if (!get_all_messages(all_entries)) {
-        LOKI_LOG(error) << "could not retrieve entries from the database\n";
+        LOKI_LOG(error, "Could not retrieve entries from the database");
         return;
     }
 
@@ -990,7 +967,7 @@ void ServiceNode::bootstrap_swarms(
     /// See what pubkeys we have
     std::unordered_map<std::string, swarm_id_t> cache;
 
-    LOKI_LOG(debug) << "we have " << all_entries.size() << " messages\n";
+    LOKI_LOG(debug, "We have {} messages", all_entries.size());
 
     std::unordered_map<swarm_id_t, std::vector<Item>> to_relay;
 
@@ -1019,7 +996,7 @@ void ServiceNode::bootstrap_swarms(
         }
     }
 
-    LOKI_LOG(trace) << "Bootstrapping " << to_relay.size() << " swarms";
+    LOKI_LOG(trace, "Bootstrapping {} swarms", to_relay.size());
 
     for (const auto& kv : to_relay) {
         const uint64_t swarm_id = kv.first;
@@ -1054,7 +1031,7 @@ void ServiceNode::relay_messages(
     }
 #endif
 
-    LOKI_LOG(info) << "serialized batches: " << data.size();
+    LOKI_LOG(info, "Serialised batches: {}", data.size());
     for (const sn_record_t& sn : snodes) {
         for (const std::shared_ptr<request_t>& batch : batches) {
             send_sn_request(batch, sn);
@@ -1081,7 +1058,7 @@ int ServiceNode::get_curr_pow_difficulty() const {
 
 bool ServiceNode::get_all_messages(std::vector<Item>& all_entries) const {
 
-    LOKI_LOG(trace) << "get all messages";
+    LOKI_LOG(trace, "Get all messages");
 
     return db_->retrieve("", all_entries, "");
 }
@@ -1094,10 +1071,9 @@ void ServiceNode::process_push_batch(const std::string& blob) {
 
     std::vector<message_t> messages = deserialize_messages(blob);
 
-    LOKI_LOG(trace) << "saving all: begin";
+    LOKI_LOG(trace, "Saving all: begin");
 
-    LOKI_LOG(debug) << "got " << messages.size()
-                    << " messages from peers, size: " << blob.size();
+    LOKI_LOG(debug, "Got {} messages from peers, size: {}", messages.size(), blob.size());
 
 #ifndef DISABLE_POW
     const auto it = std::remove_if(
@@ -1106,8 +1082,7 @@ void ServiceNode::process_push_batch(const std::string& blob) {
         });
     messages.erase(it, messages.end());
     if (it != messages.end()) {
-        LOKI_LOG(warning)
-            << "Some of the batch messages were removed due to incorrect PoW";
+        LOKI_LOG(warn, "Some of the batch messages were removed due to incorrect PoW");
     }
 #endif
 
@@ -1125,12 +1100,12 @@ void ServiceNode::process_push_batch(const std::string& blob) {
 
     save_bulk(items);
 
-    LOKI_LOG(trace) << "saving all: end";
+    LOKI_LOG(trace, "Saving all: end");
 }
 
 bool ServiceNode::is_pubkey_for_us(const std::string& pk) const {
     if (!swarm_) {
-        LOKI_LOG(error) << "swarm data missing";
+        LOKI_LOG(error, "Swarm data missing");
         return false;
     }
     return swarm_->is_pubkey_for_us(pk);
@@ -1139,7 +1114,7 @@ bool ServiceNode::is_pubkey_for_us(const std::string& pk) const {
 std::vector<sn_record_t> ServiceNode::get_snodes_by_pk(const std::string& pk) {
 
     if (!swarm_) {
-        LOKI_LOG(error) << "swarm data missing";
+        LOKI_LOG(error, "Swarm data missing");
         return {};
     }
 
@@ -1155,7 +1130,7 @@ std::vector<sn_record_t> ServiceNode::get_snodes_by_pk(const std::string& pk) {
             return si.snodes;
     }
 
-    LOKI_LOG(fatal) << "Something went wrong in get_snodes_by_pk";
+    LOKI_LOG(error, "Something went wrong in get_snodes_by_pk");
 
     return {};
 }
@@ -1164,7 +1139,7 @@ bool ServiceNode::is_snode_address_known(const std::string& sn_address) {
 
     // TODO: need more robust handling of uninitialized swarm_
     if (!swarm_) {
-        LOKI_LOG(error) << "swarm data missing";
+        LOKI_LOG(error, "Swarm data missing");
         return {};
     }
 
